+++ src/components/Cart/Cart.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Cart = () => {
+  return <div className="mt-10 text-center text-6xl font-black">Cart</div>;
+};
+
+export default Cart;